# applications/v4/api/openapi.yaml
openapi: 3.0.3
info:
  title: Weekly Planning API
  version: 1.0.0
  description: API contract for accounts, persons, projects, weeks, and weekly planning.
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
...
  /weeks:
    get:
      tags: [Weeks]
      operationId: listWeeks
      x-spring-paginated: true
      parameters:
        - $ref: '#/components/parameters/PageQueryParam'
        - $ref: '#/components/parameters/SizeQueryParam'
        - $ref: '#/components/parameters/SortQueryParam'
        - $ref: '#/components/parameters/WeekStatusQueryParam'
        - $ref: '#/components/parameters/WeekStartFromQueryParam'
        - $ref: '#/components/parameters/WeekStartToQueryParam'
      responses:
        '200': { $ref: '#/components/responses/WeekList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    post:
      tags: [Weeks]
      operationId: createWeek
      requestBody:
        $ref: '#/components/requestBodies/CreateWeekRequestBody'
      responses:
        '201': { $ref: '#/components/responses/WeekCreated' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /weeks/{weekId}:
    get:
      tags: [Weeks]
      operationId: getWeekById
      parameters:
        - $ref: '#/components/parameters/WeekIdPathParam'
      responses:
        '200': { $ref: '#/components/responses/WeekDetails' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    put:
      tags: [Weeks]
      operationId: updateWeekById
      description: ADMIN-only. When a week is set to COMPLETED, planning write operations are rejected.
      parameters:
        - $ref: '#/components/parameters/WeekIdPathParam'
      requestBody:
        $ref: '#/components/requestBodies/UpdateWeekRequestBody'
      responses:
        '200': { $ref: '#/components/responses/WeekDetails' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    delete:
      tags: [Weeks]
      operationId: deleteWeekById
      parameters:
        - $ref: '#/components/parameters/WeekIdPathParam'
      responses:
        '204': { $ref: '#/components/responses/NoContent' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /weeks/{weekId}/planning:
    get:
      tags: [Planning]
      operationId: getWeekPlanningBoard
      description: VIEWER users only receive planning cards linked to their own person.
      parameters:
        - $ref: '#/components/parameters/WeekIdPathParam'
        - $ref: '#/components/parameters/PlanningPersonIdQueryParam'
        - $ref: '#/components/parameters/IncludeEmptyDaysQueryParam'
      responses:
        '200': { $ref: '#/components/responses/WeekPlanningBoardResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /weeks/{weekId}/day-plans:
    get:
      tags: [Planning]
      operationId: listWeekDayPlans
      parameters:
        - $ref: '#/components/parameters/WeekIdPathParam'
        - $ref: '#/components/parameters/IncludeAssignmentsQueryParam'
      responses:
        '200': { $ref: '#/components/responses/DayPlanListResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /weeks/{weekId}/day-plans/{dayPlanId}:
    get:
      tags: [Planning]
      operationId: getWeekDayPlanById
      parameters:
        - $ref: '#/components/parameters/WeekIdPathParam'
        - $ref: '#/components/parameters/DayPlanIdPathParam'
      responses:
        '200': { $ref: '#/components/responses/DayPlanDetailsResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /weeks/{weekId}/day-plans/{dayPlanId}/day-persons:
    post:
      tags: [Planning]
      operationId: addPersonToDayPlan
      description: ADMIN-only. Rejects duplicate person assignment for the same day and read-only completed weeks.
      parameters:
        - $ref: '#/components/parameters/WeekIdPathParam'
        - $ref: '#/components/parameters/DayPlanIdPathParam'
      requestBody:
        $ref: '#/components/requestBodies/AddPersonToDayRequestBody'
      responses:
        '201': { $ref: '#/components/responses/DayPersonAdded' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/PlanningConflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /weeks/{weekId}/day-plans/{dayPlanId}/day-persons/{dayPersonId}:
    delete:
      tags: [Planning]
      operationId: removePersonFromDayPlan
      parameters:
        - $ref: '#/components/parameters/WeekIdPathParam'
        - $ref: '#/components/parameters/DayPlanIdPathParam'
        - $ref: '#/components/parameters/DayPersonIdPathParam'
      responses:
        '204': { $ref: '#/components/responses/NoContent' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/PlanningConflict' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /weeks/{weekId}/day-plans/{dayPlanId}/day-persons/{dayPersonId}/day-person-projects:
    post:
      tags: [Planning]
      operationId: addProjectToDayPerson
      description: ADMIN-only. Rejects duplicate project assignment for the same person/day and read-only completed weeks.
      parameters:
        - $ref: '#/components/parameters/WeekIdPathParam'
        - $ref: '#/components/parameters/DayPlanIdPathParam'
        - $ref: '#/components/parameters/DayPersonIdPathParam'
      requestBody:
        $ref: '#/components/requestBodies/AddProjectToDayPersonRequestBody'
      responses:
        '201': { $ref: '#/components/responses/DayPersonProjectAdded' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/PlanningConflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /weeks/{weekId}/day-plans/{dayPlanId}/day-persons/{dayPersonId}/day-person-projects/{dayPersonProjectId}:
    delete:
      tags: [Planning]
      operationId: removeProjectFromDayPerson
      parameters:
        - $ref: '#/components/parameters/WeekIdPathParam'
        - $ref: '#/components/parameters/DayPlanIdPathParam'
        - $ref: '#/components/parameters/DayPersonIdPathParam'
        - $ref: '#/components/parameters/DayPersonProjectIdPathParam'
      responses:
        '204': { $ref: '#/components/responses/NoContent' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/PlanningConflict' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /weeks/{weekId}/day-plans/{dayPlanId}/day-persons/{dayPersonId}/day-person-projects/{dayPersonProjectId}/tasks:
    post:
      tags: [Planning]
      operationId: addTaskToDayPersonProject
      description: ADMIN-only. Task description is mandatory and operation is rejected for completed weeks.
      parameters:
        - $ref: '#/components/parameters/WeekIdPathParam'
        - $ref: '#/components/parameters/DayPlanIdPathParam'
        - $ref: '#/components/parameters/DayPersonIdPathParam'
        - $ref: '#/components/parameters/DayPersonProjectIdPathParam'
      requestBody:
        $ref: '#/components/requestBodies/AddTaskRequestBody'
      responses:
        '201': { $ref: '#/components/responses/TaskAdded' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/PlanningConflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /weeks/{weekId}/day-plans/{dayPlanId}/day-persons/{dayPersonId}/day-person-projects/{dayPersonProjectId}/tasks/{taskId}:
    delete:
      tags: [Planning]
      operationId: removeTaskFromDayPersonProject
      parameters:
        - $ref: '#/components/parameters/WeekIdPathParam'
        - $ref: '#/components/parameters/DayPlanIdPathParam'
        - $ref: '#/components/parameters/DayPersonIdPathParam'
        - $ref: '#/components/parameters/DayPersonProjectIdPathParam'
        - $ref: '#/components/parameters/TaskIdPathParam'
      responses:
        '204': { $ref: '#/components/responses/NoContent' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/PlanningConflict' }
        '500': { $ref: '#/components/responses/InternalServerError' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    LocationHeader:
      description: URI of the created resource.
      schema:
        type: string
        format: uri
  parameters:
...
    WeekIdPathParam:
      name: weekId
      in: path
      required: true
      schema: { type: string, format: uuid }
    DayPlanIdPathParam:
      name: dayPlanId
      in: path
      required: true
      schema: { type: string, format: uuid }
    DayPersonIdPathParam:
      name: dayPersonId
      in: path
      required: true
      schema: { type: string, format: uuid }
    DayPersonProjectIdPathParam:
      name: dayPersonProjectId
      in: path
      required: true
      schema: { type: string, format: uuid }
    TaskIdPathParam:
      name: taskId
      in: path
      required: true
      schema: { type: string, format: uuid }
    PageQueryParam:
      name: page
      in: query
      schema: { type: integer, format: int32, minimum: 0, default: 0 }
    SizeQueryParam:
      name: size
      in: query
      schema: { type: integer, format: int32, minimum: 1, maximum: 100, default: 20 }
    SortQueryParam:
      name: sort
      in: query
      schema:
        type: array
        items: { type: string }
      style: form
      explode: true
...
    WeekStatusQueryParam:
      name: status
      in: query
      schema: { $ref: '#/components/schemas/WeekStatus' }
    WeekStartFromQueryParam:
      name: weekStartFrom
      in: query
      schema: { type: string, format: date }
    WeekStartToQueryParam:
      name: weekStartTo
      in: query
      schema: { type: string, format: date }
    PlanningPersonIdQueryParam:
      name: personId
      in: query
      schema: { type: string, format: uuid }
    IncludeEmptyDaysQueryParam:
      name: includeEmptyDays
      in: query
      schema: { type: boolean, default: true }
    IncludeAssignmentsQueryParam:
      name: includeAssignments
      in: query
      schema: { type: boolean, default: true }
  requestBodies:
...
    CreateWeekRequestBody:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/CreateWeekRequest' }
    UpdateWeekRequestBody:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/UpdateWeekRequest' }
    AddPersonToDayRequestBody:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AddPersonToDayRequest' }
    AddProjectToDayPersonRequestBody:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AddProjectToDayPersonRequest' }
    AddTaskRequestBody:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AddTaskRequest' }
  responses:
    NoContent:
      description: No content.
...
    WeekList:
      description: Paginated weeks.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/WeekPage' }
    WeekDetails:
      description: Week details.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/WeekDetail' }
    WeekCreated:
      description: Week created.
      headers:
        Location: { $ref: '#/components/headers/LocationHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/WeekDetail' }
    WeekPlanningBoardResponse:
      description: Planning board.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/WeekPlanningBoard' }
    DayPlanListResponse:
      description: Week day plans.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/DayPlanCollection' }
    DayPlanDetailsResponse:
      description: Day plan details.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/DayPlan' }
    DayPersonAdded:
      description: Day person assignment created.
      headers:
        Location: { $ref: '#/components/headers/LocationHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/DayPerson' }
    DayPersonProjectAdded:
      description: Day person project assignment created.
      headers:
        Location: { $ref: '#/components/headers/LocationHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/DayPersonProject' }
    TaskAdded:
      description: Task created.
      headers:
        Location: { $ref: '#/components/headers/LocationHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Task' }
    BadRequest:
      description: Bad request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Unauthorized.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Forbidden.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Conflict.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    PlanningConflict:
      description: Conflict with planning invariants or completed week.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ValidationError:
      description: Validation error.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    InternalServerError:
      description: Internal server error.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    PersonStatus:
      type: string
      enum: [ACTIVE, INACTIVE]
    ProjectStatus:
      type: string
      enum: [ACTIVE, INACTIVE]
    WeekStatus:
      type: string
      enum: [PLANNED, COMPLETED]
    DayOfWeek:
      type: string
      enum: [MONDAY, TUESDAY, WEDNESDAY, THURSDAY, FRIDAY, SATURDAY, SUNDAY]
    SortDirection:
      type: string
      enum: [asc, desc]
...
    CreateWeekRequest:
      type: object
      required: [weekStart, weekEnd]
      properties:
        weekStart:
          type: string
          format: date
        weekEnd:
          type: string
          format: date
        status:
          allOf:
            - $ref: '#/components/schemas/WeekStatus'
          default: PLANNED
    UpdateWeekRequest:
      type: object
      required: [weekStart, weekEnd, status]
      properties:
        weekStart:
          type: string
          format: date
        weekEnd:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/WeekStatus'
    WeekSummary:
      type: object
      required: [weekId, weekStart, weekEnd, status]
      properties:
        weekId:
          type: string
          format: uuid
        weekStart:
          type: string
          format: date
        weekEnd:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/WeekStatus'
    WeekDetail:
      allOf:
        - $ref: '#/components/schemas/WeekSummary'
        - type: object
          required: [createdAt, updatedAt]
          properties:
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    WeekPage:
      type: object
      required: [content, page]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/WeekSummary'
        page:
          $ref: '#/components/schemas/PageMetadata'
    WeekPlanningBoard:
      type: object
      required: [week, dayPlans]
      properties:
        week:
          $ref: '#/components/schemas/WeekDetail'
        dayPlans:
          type: array
          items:
            $ref: '#/components/schemas/DayPlan'
    DayPlanCollection:
      type: object
      required: [weekId, dayPlans]
      properties:
        weekId:
          type: string
          format: uuid
        dayPlans:
          type: array
          items:
            $ref: '#/components/schemas/DayPlan'
    DayPlan:
      type: object
      required: [dayPlanId, weekId, date, dayOfWeek, dayPersons]
      properties:
        dayPlanId:
          type: string
          format: uuid
        weekId:
          type: string
          format: uuid
        date:
          type: string
          format: date
        dayOfWeek:
          $ref: '#/components/schemas/DayOfWeek'
        dayPersons:
          type: array
          items:
            $ref: '#/components/schemas/DayPerson'
    DayPerson:
      type: object
      required: [dayPersonId, person, dayPersonProjects]
      properties:
        dayPersonId:
          type: string
          format: uuid
        person:
          $ref: '#/components/schemas/PersonReference'
        dayPersonProjects:
          type: array
          items:
            $ref: '#/components/schemas/DayPersonProject'
    DayPersonProject:
      type: object
      required: [dayPersonProjectId, project, tasks]
      properties:
        dayPersonProjectId:
          type: string
          format: uuid
        project:
          $ref: '#/components/schemas/ProjectReference'
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
    Task:
      type: object
      required: [taskId, description, createdAt, updatedAt]
      properties:
        taskId:
          type: string
          format: uuid
        description:
          type: string
          minLength: 1
          maxLength: 2000
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PersonReference:
      type: object
      required: [personId, name, email, status]
      properties:
        personId:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 120
        email:
          type: string
          format: email
          maxLength: 320
        status:
          $ref: '#/components/schemas/PersonStatus'
    ProjectReference:
      type: object
      required: [projectId, name, code, status]
      properties:
        projectId:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 120
        code:
          type: string
          maxLength: 40
        status:
          $ref: '#/components/schemas/ProjectStatus'
    AddPersonToDayRequest:
      type: object
      required: [personId]
      properties:
        personId:
          type: string
          format: uuid
    AddProjectToDayPersonRequest:
      type: object
      required: [projectId]
      properties:
        projectId:
          type: string
          format: uuid
    AddTaskRequest:
      type: object
      required: [description]
      properties:
        description:
          type: string
          minLength: 1
          maxLength: 2000
    PageMetadata:
      type: object
      required: [page, size, totalElements, totalPages]
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          format: int32
          minimum: 0
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortOrder'
    SortOrder:
      type: object
      required: [property, direction]
      properties:
        property:
          type: string
        direction:
          $ref: '#/components/schemas/SortDirection'
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
        rejectedValue:
          type: string
          nullable: true
    ErrorResponse:
      type: object
      required: [timestamp, status, error, code, message, path]
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        error:
          type: string
        code:
          type: string
        message:
          type: string
        path:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'